Replaced 5V regulator with cold-resistant model
The MC78M05CDTG ON-Semiconductor regulator was replaced with the
MC78M05BDTG. The minimum operating temperature of former is 0°C and the
latter is -45°C. There have been issues connecting to the NMX via
Bluetooth when it is very cold, so this change is being made in an
attempt to address that issue.

The CAM files have not changed at all, but their names have been updated
to Rev 3.