Increased copper pour isolation, fixed DC jack holes
On the first prototype boards, a manufacturing defect caused the ground
plane to make contact with one of the via pads, causing a very difficult
to identify problem with BLE connectivity. The clearance between both
GND and 5V copper pours and traces, vias, and pads has been increased to
avoid this in the future.

The drill holes for the DC jack were just a bit too small, so they have
been modified to allow for easy insertion of the component.